Add files via upload
Arduino Based Autonomous GPS Vehicle
My robot is a mobile platform of any kind of exploration robot so that it can navigates through designated waypoints while trying to avoid obstacles. And finally comes back to the starting point or another place. To minimize the time and money for building a new mobile platform, it is based on commonly used scale RC model car with some modification. 
The goal of the project was to create a vehicle that can autonomously navigate through a series of waypoints (GPS coordinates) while avoiding any obstacles it encounters along the way.
/*
 ------------------Pin Configation For modules
 Pins POWERS
 GPS- 3.3V
 Compass- SDA- & SCL- {ANALOG PINS} 5.0V
 Ping - TRIGER- & ECHO - 5.0V
 Drive MOtor- Motor No. 2 {Motor Shield} EXTERNAL POWER
REQURIED FOR BETTER WORKING FOR BOTH MOTOR
 Turn Motor- Motor No. 4 {Motor Shield}
*/